Add input variable stickiness_enabled
It set to true, the target group will forward a client to the same
target.
Normally, we want a browsing sessiont work with the same target to
ensure consistent experience when website content on a target is being
modified by continous deployment.
The default value, though is false to be consitent with previous
versions of the module.