Fix dependency on certificate
The SSL listener is attached to a certificate. Or vice versa, you might
say. By default Terraform doesn't do a job of orchestrating changes
well. When you update DNS records, Terraform will have to re-create the
certificate, but it wants to delete the certificate first w/o removing
it from the listener. The delete operation fails. To solve this problem
the certificate resouce now has a lifecycle rule
`create_before_destroy`.

The second problem. The listener must be attached to a valid
certificate. When you just create the certificate, it takes time for AWS
to validate it. To ensure the certificate is valid, the SSL listener has
a dependency on `aws_acm_certificate_validation.website`.

The `tests/test_update_dns.py` test verifies the correct behavior
without failures.