Use A alias again (#17)
In the version 2.* I added support of a custom AWS provider for DNS
changes. It was needed if a zone was hosted in a separare AWS account or
required specific provider configuration.

That change made me rework the DNS part of the module. I mistakenly
assumed the module can't use Amazon alias feature anymore. So, I changed
how DNS records were created. The apex A record (foo.com) was created
using public IP addresses on the load balancer. The module got them from
a data source `aws_network_interface`. That method has two critical
flaws.

- Even when the load balancer is created the network interfaces don't
immediately show up in the data source, so it causes apply failures.
- The network interfaces change! I saw somewhere in docs or forums that
once created, the load balancer nodes do not change. Well, maybe it's
true for NLBs, but not for ALBs (proven by a downtime).

The same time, I realised that my assumption about impossibility of using
Amazon's A record alias was wrong. So I tested and proved it can be
done.

This PR brings back the old proven way of creating the DNS records as it
was done in the 1.* versions.