Add new variable asg_name
In some cases a user needs to provide a fixed autoscaling group name. The asg_name variable will be used as a launch_template name and the ASG name will use the launchtemaplate attribute name. That way the module preserves ASG dependency on the launch template (needed for zero-downtime migrations) and provides the user with ability to set the ASG name.